requirements.txt
Flask
Werkzeug
pyjwt
python-dotenv